+++ Dreamline/Dreamline/Render/BarrierNode.swift
@@ -1,179 +1,13 @@
 //
-//  GameRenderer.swift
+//  BarrierNode.swift
 //  Dreamline
 //
-//  Created by BeeDream on 3/7/18.
+//  Created by BeeDream on 3/8/18.
 //  Copyright © 2018 Team BeeDream. All rights reserved.
 //
 
 import Foundation
-import UIKit
 import SpriteKit
-
-protocol GameRenderer {
-    func render(state: ModelState, config: GameConfig, events: [Event])
-}
-
-// TODO: move to own file
-class DebugRenderer: SKNode, GameRenderer {
-    // @NOTE: I think Xcode has smoothing on all
-    // of its UI, or everything has artificially
-    // smooth animation so it feels soft.
-    
-    var playerNode: SKNode
-    var barrierCache: [Int: BarrierNode]
-    var cachedFrame: CGRect
-    
-    init(frame: CGRect) {
-        self.cachedFrame = frame
-        self.barrierCache = [Int: BarrierNode]()
-        let radius = CGFloat(0.2) * (320.0) / 4.0 // @HARDCODED
-        let playerGraphic = SKShapeNode(circleOfRadius: radius)
-        playerGraphic.lineWidth = 0
-        playerGraphic.fillColor = SKColor.red
-        self.playerNode = playerGraphic
-        super.init() // this is super awkward
-        self.awake()
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func awake() {
-        self.playerNode.position = self.point(x: -9999.9, y: -9999.9) // haha
-        addChild(self.playerNode)
-    }
-    
-    func render(state: ModelState, config: GameConfig, events: [Event]) {
-        //print("rendering")
-        
-        // 1. loop through event queue
-        for event in events {
-            switch (event) {
-            case .barrierAdded(let barrier):
-                self.cacheBarrier(barrier, layout: state.boardLayout)
-            case .barrierDestroyed(let barrierId):
-                self.deleteBarrier(barrierId)
-            case .barrierPass(let id):
-                self.passBarrier(id)
-            case .barrierHit(let id):
-                self.hitBarrier(id)
-            default: break
-            }
-        }
-        
-        // 2. update all barriers
-        for barrierData in state.boardState.barriers {
-            let barrierNode = self.barrierCache[barrierData.id]!
-            barrierNode.position.y = point(x: 0.0, y: barrierData.position).y
-        }
-        
-        // update player
-        let position = state.positioner.getPosition(state: state.positionerState, config: config)
-        let offset = position.offset * state.boardLayout.laneOffset
-        self.playerNode.position = point(x: offset, y: state.boardLayout.playerPosition)
-        //self.playerNode.fillColor = position.withinTolerance ? SKColor.green : SKColor.white
-    }
-//}
-//extension DebugRender {
-    private func cacheBarrier(_ barrier: Barrier, layout: BoardLayout) {
-        let node = BarrierNode(layout: layout, width: Double(self.cachedFrame.width))
-        node.drawOnce(barrier: barrier)
-        addChild(node)
-        self.barrierCache[barrier.id] = node
-    }
-    
-    private func deleteBarrier(_ id: Int) {
-        let node = self.barrierCache[id]!
-        node.removeFromParent()
-        //node = nil
-        self.barrierCache[id] = nil
-    }
-    
-    private func passBarrier(_ barrierId: Int) {
-        let node = self.barrierCache[barrierId]!
-        node.run(SKAction.sequence([
-            SKAction.run { node.makeColor(.white) },
-            SKAction.wait(forDuration: 0.05),
-            SKAction.run { node.makeColor(.green) },
-            SKAction.fadeAlpha(to: 0.0, duration: 0.3)]))
-    }
-    
-    private func hitBarrier(_ barrierId: Int) {
-        let node = self.barrierCache[barrierId]!
-        node.run(SKAction.sequence([
-            SKAction.run { node.makeColor(.white) },
-            SKAction.wait(forDuration: 0.05),
-            SKAction.run { node.makeColor(.red) },
-            SKAction.fadeAlpha(to: 0.0, duration: 0.5)]))
-    }
-    
-    // this is weird
-    func drawLayoutLines(layout: BoardLayout) {
-        // spawn line
-        addChild(createLine(
-            from: point(x: -1.0, y: layout.spawnPosition),
-            to: point(x: 1.0, y: layout.spawnPosition),
-            color: SKColor.gray,
-            width: 1.0))
-        
-        // destroy line
-        addChild(createLine(
-            from: point(x: -1.0, y: layout.destroyPosition),
-            to: point(x: 1.0, y: layout.destroyPosition),
-            color: SKColor.gray,
-            width: 1.0))
-        
-        // player line
-        addChild(createLine(
-            from: point(x: -1.0, y: layout.playerPosition),
-            to: point(x: 1.0, y: layout.playerPosition),
-            color: SKColor.gray,
-            width: 1.0))
-        
-        // lane lines
-        addChild(createLine(
-            from: point(x: -layout.laneOffset, y: layout.playerPosition),
-            to: point(x: -layout.laneOffset, y: -1.0),
-            color: SKColor.gray,
-            width: 1.0))
-        addChild(createLine(
-            from: point(x: 0.0, y: layout.playerPosition),
-            to: point(x: 0.0, y: -1.0),
-            color: SKColor.gray,
-            width: 1.0))
-        addChild(createLine(
-            from: point(x: layout.laneOffset, y: layout.playerPosition),
-            to: point(x: layout.laneOffset, y: -1.0),
-            color: SKColor.gray,
-            width: 1.0))
-    }
-    
-    private func point(x: Double, y: Double) -> CGPoint {
-        return point(x: CGFloat(x), y: CGFloat(y))
-    }
-    
-    private func point(x: CGFloat, y: CGFloat) -> CGPoint {
-        let convertedX = self.cachedFrame.midX + (x * self.cachedFrame.width / 2.0)
-        let convertedY = self.cachedFrame.midY - (y * self.cachedFrame.height / 2.0)
-        return CGPoint(x: convertedX, y: convertedY)
-    }
-    
-    private func createLine(from: CGPoint,
-                            to: CGPoint,
-                            color: SKColor,
-                            width: CGFloat = 2.0) -> SKShapeNode {
-        let node = SKShapeNode()
-        let path = UIBezierPath()
-        path.move(to: from)
-        path.addLine(to: to)
-        node.path = path.cgPath
-        node.strokeColor = color
-        node.lineWidth = width
-        return node
-    }
-}
 
 class BarrierNode: SKNode {
